decimal values for delay time in samples; deactivation of 2 GUI elements

+++ Source/DelayProcessor.cpp
@@ -87,7 +87,7 @@ void DelayProcessor::setDelayTime(double delayTimeInSeconds, double sampleRate) 
     currentDelayTime = clampedDelayTime; // in seconds
 }
 
-void DelayProcessor::setDelayTimeInSamples(int delayTimeInSamples, double sampleRate) {
+void DelayProcessor::setDelayTimeInSamples(float delayTimeInSamples, double sampleRate) {
     // double clampedDelayTimeInSamples = juce::jlimit(0.0, maxDelayTime * sampleRate, delayTimeInSamples * sampleRate);
     delayLine.setDelay(delayTimeInSamples);
     currentDelayInSamples = delayTimeInSamples;
